requirements.txt
huggingface_hub>=0.18.0
llama-cpp-python>=0.2.20
fastapi
uvicorn
pydantic
numpy
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.3.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
python-dotenv>=1.0.0
